feat: 대댓글이 있는 부모 댓글의 경우 소프트 삭제 (#55)

+++ src/main/java/com/team/buddyya/feed/service/CommentService.java
@@ -76,7 +76,13 @@ public class CommentService {
         Feed feed = findFeedByFeedId(feedId);
         Comment comment = findCommentByCommentId(commentId);
         validateCommentOwner(studentInfo.id(), comment);
-        commentRepository.delete(comment);
+        boolean hasChild = !comment.getChildren().isEmpty();
+        if (hasChild) {
+            comment.updateIsDeleted(true);
+            commentRepository.save(comment);
+        } else {
+            commentRepository.delete(comment);
+        }
     }
 
     private void validateCommentOwner(Long studentId, Comment comment) {
